refactor(settings): split terminal behavior into its own section
Move Terminal scrollback, scroll minimap, select-to-copy, and plain/bracketed
paste out of Appearance into a new Terminal section, sitting with the
session/connection cluster (Terminal / SFTP / Port Forwarding / Hosts).
Appearance keeps scale, language, theme, and automation.

i18n keys moved to a settings.terminal namespace with nav label + keywords
across en/fr/ru; quick-toggle labels under toggleDefs left untouched.
