+++ src/components/settings/sections/AppearanceSection.tsx
@@ -4,14 +4,9 @@ import { useThemeStore } from "@/stores/themeStore";
 import { usePluginStore } from "@/stores/pluginStore";
 import { BUILT_IN_THEMES } from "@/themes/presets";
 import { useUIStore } from "@/stores/uiStore";
-import { useTerminalSettingsStore } from "@/stores/terminalSettingsStore";
-import { TOGGLE_DEFS, useToggle } from "@/stores/toggleSettingsStore";
-import { DEFAULT_SCROLLBACK_LINES, MAX_SCROLLBACK_LINES, MIN_SCROLLBACK_LINES } from "@/stores/terminalSettingsUtils";
 import { FormSelect } from "@/components/shared/FormSelect";
-import { Toggle } from "@/components/shared/Toggle";
 import type { AppTheme } from "@/themes/types";
 import ScaleSection from "./ScaleSection";
-import { DirtyDot, ResetButton } from "./shared";
 import { useLocaleStore, SUPPORTED_LOCALES } from "@/stores/localeStore";
 
 function downloadJson(filename: string, data: unknown) {
@@ -35,11 +30,6 @@ export default function AppearanceSection() {
   const { activeThemeId, customThemes, setTheme, deleteCustomTheme } = useThemeStore();
   const { openThemeCreator, openThemeImportExport } = useUIStore();
   const pluginThemeMap = usePluginStore((s) => s.pluginThemes);
-  const [scrollMinimapEnabled, setScrollMinimapEnabled] = useToggle("scroll-minimap");
-  const [selectToCopy, setSelectToCopy] = useToggle("select-to-copy");
-  const [ignoreBracketedPaste, setIgnoreBracketedPaste] = useToggle("ignore-bracketed-paste");
-  const scrollbackLines = useTerminalSettingsStore((s) => s.scrollbackLines);
-  const setScrollbackLines = useTerminalSettingsStore((s) => s.setScrollbackLines);
 
   const { t } = useTranslation();
   const locale = useLocaleStore((s) => s.locale);
@@ -47,9 +37,6 @@ export default function AppearanceSection() {
 
   const pluginThemes: AppTheme[] = [...pluginThemeMap.values()].map((theme) => ({ ...theme, builtIn: true }));
   const allThemes = [...BUILT_IN_THEMES, ...customThemes, ...pluginThemes];
-  const scrollbackOptions = [1_000, 10_000, 50_000, 100_000, 250_000]
-    .filter((value) => value >= MIN_SCROLLBACK_LINES && value <= MAX_SCROLLBACK_LINES)
-    .map((value) => ({ value: String(value), label: t("settings.appearance.scrollback.option", { count: value }) }));
 
   const handleDelete = (id: string) => {
     deleteCustomTheme(id);
@@ -74,71 +61,6 @@ export default function AppearanceSection() {
               options={SUPPORTED_LOCALES}
               onChange={(value) => setLocale(value as typeof locale)}
             />
-          </div>
-        </div>
-        <div className="group mt-4 rounded-xl bg-(--t-bg-card) border border-(--t-border) p-4 flex items-center justify-between gap-4">
-          <div>
-            <div className="text-sm font-medium text-(--t-text-primary)">{t("settings.appearance.scrollback.title")}</div>
-            <div className="text-xs mt-1 text-(--t-text-dim)">
-              {t("settings.appearance.scrollback.desc")}
-            </div>
-          </div>
-          <div className="flex items-center gap-2 shrink-0">
-            {scrollbackLines !== DEFAULT_SCROLLBACK_LINES && (
-              <ResetButton onReset={() => setScrollbackLines(DEFAULT_SCROLLBACK_LINES)} />
-            )}
-            {scrollbackLines !== DEFAULT_SCROLLBACK_LINES && <DirtyDot />}
-            <FormSelect
-              className="w-44 shrink-0"
-              value={String(scrollbackLines)}
-              options={scrollbackOptions}
-              onChange={(value) => setScrollbackLines(Number(value))}
-            />
-          </div>
-        </div>
-        <div className="group mt-4 rounded-xl bg-(--t-bg-card) border border-(--t-border) p-4 flex items-center justify-between gap-4">
-          <div>
-            <div className="text-sm font-medium text-(--t-text-primary)">{t("settings.appearance.minimap.title")}</div>
-            <div className="text-xs mt-1 text-(--t-text-dim)">
-              {t("settings.appearance.minimap.desc")}
-            </div>
-          </div>
-          <div className="flex items-center gap-2 shrink-0">
-            {scrollMinimapEnabled !== TOGGLE_DEFS["scroll-minimap"].default && (
-              <ResetButton onReset={() => setScrollMinimapEnabled(TOGGLE_DEFS["scroll-minimap"].default)} />
-            )}
-            {scrollMinimapEnabled !== TOGGLE_DEFS["scroll-minimap"].default && <DirtyDot />}
-            <Toggle checked={scrollMinimapEnabled} onChange={setScrollMinimapEnabled} />
-          </div>
-        </div>
-        <div className="group mt-4 rounded-xl bg-(--t-bg-card) border border-(--t-border) p-4 flex items-center justify-between gap-4">
-          <div>
-            <div className="text-sm font-medium text-(--t-text-primary)">{t("settings.appearance.selectToCopy.title")}</div>
-            <div className="text-xs mt-1 text-(--t-text-dim)">
-              {t("settings.appearance.selectToCopy.desc")}
-            </div>
-          </div>
-          <div className="flex items-center gap-2 shrink-0">
-            {selectToCopy !== TOGGLE_DEFS["select-to-copy"].default && (
-              <ResetButton onReset={() => setSelectToCopy(TOGGLE_DEFS["select-to-copy"].default)} />
-            )}
-            {selectToCopy !== TOGGLE_DEFS["select-to-copy"].default && <DirtyDot />}
-            <Toggle checked={selectToCopy} onChange={setSelectToCopy} />
-          </div>
-        </div>
-        <div className="group mt-4 rounded-xl bg-(--t-bg-card) border border-(--t-border) p-4 flex items-center justify-between gap-4">
-          <div>
-            <div className="text-sm font-medium text-(--t-text-primary)">{t("settings.appearance.ignoreBracketedPaste.title")}</div>
-            <div className="text-xs mt-1 text-(--t-text-dim)">
-              {t("settings.appearance.ignoreBracketedPaste.desc")}
-            </div>
-          </div>
-          <div className="flex items-center gap-2 shrink-0">
-            {ignoreBracketedPaste !== TOGGLE_DEFS["ignore-bracketed-paste"].default && (
-              <ResetButton onReset={() => setIgnoreBracketedPaste(TOGGLE_DEFS["ignore-bracketed-paste"].default)} />
-            )}
-            {ignoreBracketedPaste !== TOGGLE_DEFS["ignore-bracketed-paste"].default && <DirtyDot />}
-            <Toggle checked={ignoreBracketedPaste} onChange={setIgnoreBracketedPaste} />
           </div>
         </div>
       </div>
