+++ src/components/settings/settingsSections.tsx
@@ -4,6 +4,7 @@ import AccountSection from "@/components/settings/sections/AccountSection";
 import SyncSection from "@/components/settings/sections/SyncSection";
 import VaultsSection from "@/components/settings/sections/VaultsSection";
 import PluginsSection from "@/components/settings/sections/PluginsSection";
+import TerminalSection from "@/components/settings/sections/TerminalSection";
 import SFTPSection from "@/components/settings/sections/SFTPSection";
 import PortForwardingSection from "@/components/settings/sections/PortForwardingSection";
 import AboutSection from "@/components/settings/sections/AboutSection";
@@ -19,6 +20,7 @@ export function renderSettingsSection(section: SettingsSection) {
     case "sync": return <SyncSection />;
     case "vaults": return <VaultsSection />;
     case "plugins": return <PluginsSection />;
+    case "terminal": return <TerminalSection />;
     case "sftp": return <SFTPSection />;
     case "portForwarding": return <PortForwardingSection />;
     case "hosts": return <HostsSection />;
