+++ src/i18n/locales/fr/settings.json
@@ -1,8 +1,7 @@
 {
   "settings": {
-    "appearance": {
-      "interface": "Interface",
-      "language": { "title": "Langue" },
+    "terminal": {
+      "heading": "Comportement du terminal",
       "scrollback": {
         "title": "Historique du terminal",
         "desc": "Lignes conservées par session de terminal. S'applique aux nouveaux terminaux.",
@@ -19,7 +18,11 @@
       "ignoreBracketedPaste": {
         "title": "Collage simple (sans collage entre crochets)",
         "desc": "Coller en texte brut sans les marqueurs \\x1b[200~ / \\x1b[201~. Corrige un affichage corrompu (ex. « 200~ ») lors d'un collage dans un prompt sudo ou une application CLI qui ne prend pas en charge le collage entre crochets."
-      },
+      }
+    },
+    "appearance": {
+      "interface": "Interface",
+      "language": { "title": "Langue" },
       "colorTheme": "Thème de couleurs",
       "import": "Importer",
       "importTitle": "Importer un ou plusieurs thèmes",
@@ -45,6 +48,7 @@
       "sync":           { "label": "Synchronisation",    "keywords": ["sauvegarde", "nuage", "gist", "synchronisation", "pro"] },
       "vaults":         { "label": "Coffres-forts",      "keywords": ["secret", "mot de passe", "stockage", "trousseau"] },
       "plugins":        { "label": "Extensions",         "keywords": ["extensions", "modules", "boutique"] },
+      "terminal":       { "label": "Terminal",           "keywords": ["historique", "collage", "crochets", "copier", "mini-carte", "tampon"] },
       "sftp":           { "label": "SFTP",               "keywords": ["fichier", "transfert", "ftp", "fichiers"] },
       "portForwarding": { "label": "Transfert de ports", "keywords": ["tunnel", "redirection", "ports", "ssh", "auto"] },
       "hosts":          { "label": "Hôtes",              "keywords": ["ping", "accessibilité", "connectivité", "statut"] },
