+++ src/i18n/locales/ru/settings.json
@@ -1,8 +1,7 @@
 {
   "settings": {
-    "appearance": {
-      "interface": "Интерфейс",
-      "language": { "title": "Язык" },
+    "terminal": {
+      "heading": "Поведение терминала",
       "scrollback": {
         "title": "Буфер прокрутки терминала",
         "desc": "Количество сохраняемых строк на сессию терминала. Применяется к новым терминалам.",
@@ -19,7 +18,11 @@
       "ignoreBracketedPaste": {
         "title": "Простая вставка (без «скобочной» вставки)",
         "desc": "Вставлять как обычный текст без маркеров \\x1b[200~ / \\x1b[201~. Исправляет искажённый ввод (например «200~») при вставке в приглашения sudo или приложения CLI, не поддерживающие «скобочную» вставку."
-      },
+      }
+    },
+    "appearance": {
+      "interface": "Интерфейс",
+      "language": { "title": "Язык" },
       "colorTheme": "Цветовая тема",
       "import": "Импорт",
       "importTitle": "Импортировать тему(-ы)",
@@ -45,6 +48,7 @@
       "sync":           { "label": "Синхронизация",       "keywords": ["резервная копия", "облако", "gist", "синхронизация", "pro"] },
       "vaults":         { "label": "Хранилища",           "keywords": ["секрет", "пароль", "хранилище", "связка ключей"] },
       "plugins":        { "label": "Плагины",             "keywords": ["расширения", "дополнения", "маркетплейс"] },
+      "terminal":       { "label": "Терминал",            "keywords": ["буфер", "вставка", "скобочная", "копирование", "мини-карта", "прокрутка"] },
       "sftp":           { "label": "SFTP",                "keywords": ["файл", "передача", "ftp", "файлы"] },
       "portForwarding": { "label": "Проброс портов",      "keywords": ["туннель", "проброс", "порты", "ssh", "авто"] },
       "hosts":          { "label": "Хосты",               "keywords": ["ping", "доступность", "подключение", "статус"] },
